# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Finance & Crisis Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List all savings goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavingsGoal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a new savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsGoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
  /goals/{id}:
    put:
      operationId: updateGoal
      tags: [goals]
      summary: Update a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsGoalInput"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── Behavioral finance ──────────────────────────────────────
...
  /financial-profile:
    get:
      operationId: getFinancialAutonomyProfile
      tags: [resilience]
      summary: Get money script, risk tolerance, and autonomy score
      responses:
        "200":
          description: Financial autonomy profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialAutonomyProfile"
    patch:
      operationId: updateFinancialAutonomyProfile
      tags: [resilience]
      summary: Update money script and risk tolerance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialAutonomyProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialAutonomyProfile"
  /vault:
    get:
      operationId: getVaultSummary
      tags: [resilience]
      summary: List document readiness and autonomy warnings
      responses:
        "200":
          description: Redacted vault document list and readiness summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultSummary"
    post:
      operationId: createVaultDocument
      tags: [resilience]
      summary: Store a vault document's encrypted payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultDocumentInput"
      responses:
        "201":
          description: Created redacted vault document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultDocument"
  /vault/{id}:
    patch:
      operationId: updateVaultDocument
      tags: [resilience]
      summary: Update vault metadata or encrypted payload
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultDocumentUpdate"
      responses:
        "200":
          description: Updated redacted vault document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultDocument"
    delete:
      operationId: deleteVaultDocument
      tags: [resilience]
      summary: Delete a vault document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /vault/{id}/verify:
    post:
      operationId: verifyVaultDocument
      tags: [resilience]
      summary: Mark a vault document as verified today
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Verified redacted vault document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultDocument"

  # ─── Hypotheses sandbox ──────────────────────────────────────
...
components:
  schemas:
...
    SavingsGoal:
      type: object
      required: [id, title, targetAmount, targetMonths, currentAmount, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        targetAmount:
          type: number
        targetMonths:
          type: number
        currentAmount:
          type: number
        createdAt:
          type: string
    SavingsGoalInput:
      type: object
      required: [title, targetAmount, targetMonths]
      properties:
        title:
          type: string
        targetAmount:
          type: number
        targetMonths:
          type: number
        currentAmount:
          type: number

    # ─── Behavioral finance ───────────────────────────────────
...
    FinancialAutonomyProfile:
      type: object
      required: [userId, moneyScriptType, riskToleranceIndex, autonomyScore, updatedAt]
      properties:
        userId:
          type: string
        moneyScriptType:
          type: string
          enum: [avoidance, worship, status, vigilance]
        riskToleranceIndex:
          type: number
        autonomyScore:
          type: number
        updatedAt:
          type: string
    FinancialAutonomyProfileUpdate:
      type: object
      properties:
        moneyScriptType:
          type: string
          enum: [avoidance, worship, status, vigilance]
        riskToleranceIndex:
          type: number
    VaultDocument:
      type: object
      required: [id, docCategory, title, createdAt]
      properties:
        id:
          type: number
        docCategory:
          type: string
          enum: [bank_account, tax_file, contract, emergency_plan]
        title:
          type: string
        lastVerifiedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    VaultDocumentInput:
      type: object
      required: [docCategory, title, encryptedPayload]
      properties:
        docCategory:
          type: string
          enum: [bank_account, tax_file, contract, emergency_plan]
        title:
          type: string
        encryptedPayload:
          type: string
          description: Opaque encrypted payload or external secure reference. Never returned by the API.
        lastVerifiedAt:
          type: string
    VaultDocumentUpdate:
      type: object
      properties:
        docCategory:
          type: string
          enum: [bank_account, tax_file, contract, emergency_plan]
        title:
          type: string
        encryptedPayload:
          type: string
        lastVerifiedAt:
          type: ["string", "null"]
    VaultCategoryStatus:
      type: object
      required: [category, present, verified, stale, lastVerifiedAt]
      properties:
        category:
          type: string
        present:
          type: boolean
        verified:
          type: boolean
        stale:
          type: boolean
        lastVerifiedAt:
          type: ["string", "null"]
    VaultSummary:
      type: object
      required: [documents, autonomyScore, warnings, categoryStatus]
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/VaultDocument"
        autonomyScore:
          type: number
        warnings:
          type: array
          items:
            type: string
        categoryStatus:
          type: array
          items:
            $ref: "#/components/schemas/VaultCategoryStatus"

    # ─── Hypotheses sandbox ───────────────────────────────────
...